refactor: 단계 카드 UI 공통 컴포넌트 StepCardList로 분리 및 SectionTable 적용(#247)
